Move portable/util.py into extension/export_util
`portable/util.py` is used by multiple example models including llm models. Since we are moving llm models' export libraries into `extension/llm/export`, it makes sense to move `portable/util.py`
into extension as well.

This change does that.

It moves the test into `examples/models` which seems to be a better place for these unit tests since they test the export flow for different models and also rely on some code under `examples/models` like
`EagerModelFactory`.

Differential Revision: [D59306348](https://our.internmc.facebook.com/intern/diff/D59306348/)

[ghstack-poisoned]
