+++ examples/models/llama2/TARGETS
@@ -86,9 +86,9 @@ runtime.python_library(
         "//executorch/examples/models:model_base",
         "//executorch/examples/models:models",
         "//executorch/examples/models/llama2/custom_ops:custom_ops_aot_py",
-        "//executorch/examples/portable:utils",
         "//executorch/exir:lib",
         "//executorch/extension/llm/export:export_lib",
+        "//executorch/extension/export_util:export_util",
         # one definition has to be included in the user of the libarary
         # depending on what library the client wants to use
         # "//executorch/extension/pybindings:aten_lib",
